% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardiodatasets-package.R
\docType{package}
\name{CardioDataSets}
\alias{CardioDataSets}
\alias{CardioDataSets-package}
\title{CardioDataSets: A Comprehensive Collection of Cardiovascular and Heart Disease Datasets}
\description{
{
This package provides a wide variety of datasets focused on heart and cardiovascular research, covering heart disease, myocardial infarction, heart failure, stroke, ischemic heart disease, risk factors, clinical trials, and treatment outcomes.
}
}
\details{
CardioDataSets: A Comprehensive Collection of Cardiovascular and Heart Disease Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Comprehensive Collection of Cardiovascular and Heart Disease Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/cardiodatasets}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
